+++ Lib/TSCompiler/IBundler.cs
@@ -8,6 +8,8 @@ namespace Lib.TSCompiler
         RefDictionary<string, object> FilesContent { set; }
         ProjectOptions Project { set; }
         BuildResult BuildResult { set; }
+        bool BuildSourceMap { set; }
+        string? SourceMapSourceRoot { set; }
         void Build(bool compress, bool mangle, bool beautify);
     }
 }