+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Added
 
-- New bundler from Njsast purely in C#. Currently not enabled by default, but could enabled by adding `bb b -x 1`
+- New bundler from Njsast purely in C#. Currently not enabled by default, but could enabled by adding `bb b -x 1`. New bundler allows to generate source map, enable by `-g yes`. Using `--sourceRoot ".."` define what should be written in source map sourceRoot field, default is `..` which nicely works with default `dist` output directory.
 
 ## 0.97.1